% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvtv.default}
\alias{mvtv.default}
\title{Default Multivariate Total Variation Denoising Solver}
\usage{
\method{mvtv}{default}(data, y, m, mesh = NULL, n_lambda = 100L,
  ftrue = NULL, lambdas = NULL, folds = 5L, verbose = TRUE)
}
\arguments{
\item{data}{n by p matrix of inputs}

\item{y}{response column vector}

\item{m}{vector of number of mesh points per predictor}

\item{mesh}{user can supply or NULL for regularly spaced mesh, which will be returned}

\item{n_lambda}{number of logarithmically spaced tuning parameters}

\item{ftrue}{prediction target. If NULL, use observed data.}

\item{lambdas}{user can supply vector of lambdas to be solved over. If NULL,
function generates n_lambda logarithmically spaced lambdas from 0.00001*lambda_max
and lambda_max, where lambda_max is our approximation of smallest lambda where
regularization ends.}

\item{folds}{number of folds for cross-validation}

\item{verbose}{Default: true, prints out current working penalty and 
number of iters to solve.}
}
\description{
Create a mesh and find cross-validated best 
approximation to total variation denoising problem.
}
\examples{
# Approximating Bivariate Fused Lasso for Uniform Data 
## Generate Data
set.seed(117)
x <- matrix(runif(100),ncol = 2)
y <- matrix(runif(50),ncol=1)
m <- matrix(c(3,3))

## Find Total Variation Solution over range of lambdas and whole data set
mvtv_fold1 <- mvtv(x,y,m,folds=1, verbose = FALSE)

## Find 5-fold validated MVTV Model over range of lambdas
mvtv_fold5 <- mvtv(x,y,m,folds=5, verbose = FALSE)

}
