% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivarTV.R
\name{plot.mvtv}
\alias{plot.mvtv}
\title{Plotting Fitted Surface, p=1}
\usage{
\method{plot}{mvtv}(x, ..., addmesh = FALSE, adddata = TRUE,
  lambda = NULL)
}
\arguments{
\item{x}{object of class 'mvtv.'}

\item{...}{ignore.}

\item{addmesh}{If TRUE, vertical grey lines plotted along x-axis value of mesh.}

\item{adddata}{If TRUE, observed data is plotted.}

\item{lambda}{Plot at specified lambda. If NULL, plot fit at lambda with smalled cross-validated MSE.}
}
\description{
Plotting fitted values for an 'mvtv' Object
}
