% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivarTV.R
\name{plotFits}
\alias{plotFits}
\title{Plotting Fitted Surface, p=1}
\usage{
plotFits(mvtvmodel, addmesh = FALSE)
}
\arguments{
\item{mvtvmodel}{object of class 'mvtv'}

\item{addmesh}{If TRUE, plot has vertical grey lines along x-axis value of mesh}
}
\description{
Plotting fitted values for an 'mvtv' Object
}
