% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predict.mvtv}
\alias{predict.mvtv}
\title{MVTV Predict Function}
\usage{
\method{predict}{mvtv}(mvtvobject, data = NULL)
}
\arguments{
\item{mvtvobject}{object produced by mbtv.default}

\item{data}{n by p matrix of inputs}
}
\description{
Use fitted 'mvtv' object to predict new data.
}
\examples{
# Approximating Bivariate Fused Lasso for Uniform Data 
## Generate Data
set.seed(117)
x <- matrix(runif(100),ncol = 2)
y <- matrix(runif(50),ncol=1)
m <- matrix(c(3,3))

## Find 5-fold validated MBS Model over range of lambdas
mbs_fold5 <- mvtv(x,y,m,folds=5,verbose=FALSE)

# Access fitted values of training data; equivalent to mbs_fold5$fitted
fitted.values <- predict(mbs_fold5) 
newdata <- matrix( runif(50), ncol = 2) # Generate new data
newfits <- predict(mbs_fold5, newdata) # Fit new data
}
